+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.2.0-rc.1
+## 12-02-2023
+
+1. [](#new)
+    * Filter timeline by date range
+2. [](#improved)
+    * Processing overhead
+    * Qode-quality
+
 # v2.1.2
 ## 10-06-2021